Rename lasers.py to lantzlasers.py
