+++ imswitch/imcontrol/model/interfaces/lantzlasers.py
